compose: Remove below-compose-content div.
This commit focuses on further CSS cleaning of the compose box.
It removes the below-compose-content div, which served as a container
for compose_bottom_top_container and compose_bottom_bottom_container
divs. The removal of this div has minimal impact on the layout,
except for margin adjustments. To address this, a 6px margin-top
is applied to compose_bottom_top_container, while
compose_bottom_bottom_container gets a -2px margin-bottom.